Arguments[0]()
Key Concepts:

this inside arguments[0]():
When arguments[0]() is invoked, the this keyword inside the callback function refers to the arguments object because of how the function is called.

arguments.length:
The arguments object has a length property that represents the number of arguments passed to the function. In this case, it's 3 because method() was called with three arguments: callback, 1, and 2.